Assume default parameters for force-deploys
This change alters the behaviour for using `rain deploy --force [...]`:
Where an existing, or default, parameter is available then presume the
same behaviour as if the user accepted that default suggestion.

The intended use-case for this is that a template can be deployed using
rain with zero interactivity when default parameter values can be
determined.